Split LLink out to TS file
